+++ Etabs_Adapter/Etabs_Adapter.csproj
@@ -200,6 +200,7 @@
     <Compile Include="Create\_Create.cs" />
     <Compile Include="Delete\_Delete.cs" />
     <Compile Include="Read\Results\BarResults.cs" />
+    <Compile Include="Read\Results\GlobalResults.cs" />
     <Compile Include="Read\Results\NodeResults.cs" />
     <Compile Include="Read\Results\MeshResults.cs" />
     <Compile Include="Types\NextIndex.cs" />
